prep for release v2.2.0
