+++ lerna.json
@@ -1,5 +1,5 @@
 {
-  "version": "2.1.1",
+  "version": "2.2.0",
   "packages": [
     "packages/*"
   ]
